# src/swaggerDocs.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "The Biz Nation API",
    "description": "API",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "http://localhost:3000/api/disney-demo/v1"
    }
  ],
...
  "paths": {
    "/users": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Obtener todos los usuarios",
        "description": "Devuelve una lista de todos los usuarios registrados.",
        "parameters": [
          {
            "$ref": "#/components/parameters/token"
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/userResponse"
                }
              }
            }
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/users/{id}": {
      "get": {
        "tags": [
          "User"
        ],
        "summary": "Obtener un usuario por id",
        "description": "Devuelve el usuario con el id dado o un mensaje de error",
        "parameters": [
          {
            "$ref": "#/components/parameters/token"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Identificador unico del usuario",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/userResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
    "/users/{id}/": {
      "put": {
        "tags": [
          "User"
        ],
        "summary": "Actualizar un usuario",
        "description": "Actualiza el usuario indicado y lo retorna",
        "parameters": [
          {
            "$ref": "#/components/parameters/token"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Identificador unico del usuario",
            "required": true,
            "schema": {
              "type": "number"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/updateUser"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/userResponse"
                }
              }
            }
          },
          "400": {
            "$ref": "#/components/responses/BadRequest"
          },
          "401": {
            "$ref": "#/components/responses/Unauthorized"
          }
        }
      }
    },
...
  },
  "components": {
    "responses": {
      "Unauthorized": {
        "description": "(unauthorized) no ha enviado el bearer token que lo autentique"
      },
      "BadRequest": {
        "description": "(BadRequest) no cumple con los esquemas de validación de datos"
      },
...
    },
    "parameters": {
      "token": {
        "name": "Authorization",
        "in": "header",
        "description": "Token (Bearer) de autenticación.",
        "required": true,
        "schema": {
          "type": "string"
        }
      }
    },
    "schemas": {
      "userResponse": {
        "type": "object",
        "properties": {
          "id": {
            "type": "number"
          },
          "nickname": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "createdAt": {
            "type": "string"
          }
        }
      },
      "updateUser": {
        "type": "object",
        "properties": {
          "nickname": {
            "type": "string"
          },
          "email": {
            "type": "string"
          }
        }
      },
...
    }
  }
}